Increase test coverage to 100% in hypermodel
